rename revaultd module to daemon
