+++ src/daemon/client/error.rs
