+++ src/daemon/client/mod.rs
