+++ src/daemon/model.rs
