+++ src/installer/step/stakeholder.rs
@@ -9,6 +9,7 @@ use revault_tx::{
 };
 
 use crate::{
+    daemon::config,
     installer::{
         message::{self, Message},
         step::{
@@ -17,7 +18,6 @@ use crate::{
         },
         view,
     },
-    revaultd::config,
     ui::component::form,
 };
