+++ src/main.rs
@@ -8,11 +8,11 @@ extern crate serde_json;
 
 mod app;
 mod conversion;
+mod daemon;
 mod hw;
 mod installer;
 mod loader;
 mod revault;
-mod revaultd;
 mod ui;
 
 use app::{
@@ -22,10 +22,10 @@ use app::{
     App,
 };
 use conversion::Converter;
+use daemon::config::default_datadir;
 use installer::Installer;
 use loader::Loader;
 use revault::Role;
-use revaultd::config::default_datadir;
 
 enum Args {
     ConfigPath(PathBuf),
